Feature: Deleting button and field classes and unifying them
