+++ core/designsystem/src/main/java/com/iteneum/designsystem/components/Button.kt
@@ -260,3 +260,56 @@ fun LpRadioGroup(
         }
     }
 }
+
+/**
+ * Create [BadgeButton] compose for user's notifications
+ *
+ * @param modifier to modify box properties that contains: Button & Badge
+ * @param badgeNumber to modify the number of notifications shown
+ * @param onClick high order function, to define button action
+ *
+ * @author Jose Miguel Garcia Reyes
+ */
+
+@ExperimentalMaterial3Api
+@Composable
+fun BadgeButton(
+    modifier: Modifier = Modifier,
+    badgeNumber: Int,
+    onClick: () -> Unit = {}
+) {
+    Box(modifier = modifier) {
+        FilledIconButton(
+            onClick = onClick,
+            colors = IconButtonDefaults.filledIconButtonColors(MaterialTheme.colorScheme.surface)
+        ) {
+            Icon(
+                imageVector = Icons.Filled.Notifications,
+                contentDescription = "Notifications button",
+                tint = MaterialTheme.colorScheme.secondary
+            )
+        }
+        if (badgeNumber > 0) {
+            BadgedBox(
+                modifier = Modifier
+                    .padding(
+                        horizontal = 15.dp,
+                        vertical = 7.dp
+                    )
+                    .align(alignment = Alignment.BottomEnd),
+                badge = {
+                    Badge(
+                        modifier = Modifier
+                            .padding(all = 1.dp)
+                            .align(alignment = Alignment.TopEnd)
+                    ) {
+                        Text(
+                            text = badgeNumber.toString(),
+                            color = MaterialTheme.colorScheme.primary
+                        )
+                    }
+                }
+            ) {}
+        }
+    }
+}
