+++ core/designsystem/src/main/java/com/iteneum/designsystem/components/TextField.kt
@@ -1,5 +1,6 @@
 package com.iteneum.designsystem.components
 
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Error
@@ -10,6 +11,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.text.input.TextFieldValue
 import com.iteneum.designsystem.R
 
 /**
@@ -65,5 +67,103 @@ fun LpOutlinedTextFieldPassword(
         isError = showError,
         singleLine = true,
         maxLines = 1
+    )
+}
+
+/**
+ * This function creates a password OutlinedTextField
+ * @param modifier Set component modifier
+ * @param label Text label
+ * @param hint Hint message
+ * @param isValid Validate if text is valid
+ * @param supportTextError Error message
+ * @param onValueChange This parameter return the field value
+ *
+ * @author Andrés Ivan Medina Herrera
+ */
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun LpOutlinedTextFieldMail(
+    modifier: Modifier,
+    label: String,
+    hint: String,
+    isValid: Boolean,
+    supportTextError: String,
+    onValueChange: (String) -> Unit
+) {
+    var textFieldValue by remember { mutableStateOf(TextFieldValue("")) }
+
+    OutlinedTextField(
+        value = textFieldValue,
+        onValueChange = {
+            textFieldValue = it
+            onValueChange(textFieldValue.text)
+        },
+        modifier = modifier.fillMaxWidth(),
+        label = { Text(label) },
+        placeholder = { Text(hint) },
+        colors = TextFieldDefaults.outlinedTextFieldColors(
+            focusedBorderColor = MaterialTheme.colorScheme.onPrimary,
+            unfocusedBorderColor = MaterialTheme.colorScheme.onPrimary,
+            focusedLabelColor = MaterialTheme.colorScheme.onPrimary,
+            errorLabelColor = MaterialTheme.colorScheme.error,
+            errorBorderColor = MaterialTheme.colorScheme.error,
+            errorSupportingTextColor = MaterialTheme.colorScheme.error
+        ),
+        isError = isValid,
+        supportingText = {
+            if (isValid) {
+                Text(text = supportTextError)
+            } else {
+                Text(text = "")
+            }
+        },
+        singleLine = true,
+        maxLines = 1,
+        shape = MaterialTheme.shapes.small,
+        keyboardOptions = KeyboardOptions.Default.copy(
+            keyboardType = KeyboardType.Email
+        )
+    )
+}
+/**
+ * This function creates a password OutlinedTextField
+ * @param modifier Set component modifier
+ * @param label Text label
+ * @param hint Hint message
+ * @param onValueChange This parameter return the field value
+ *
+ * @author Andrés Ivan Medina Herrera
+ */
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun LpOutlinedTextFieldNumber(
+    modifier: Modifier,
+    label: String,
+    hint: String,
+    onValueChange: (String) -> Unit,
+) {
+    var textFieldValue by remember { mutableStateOf(TextFieldValue("")) }
+    OutlinedTextField(
+        value = textFieldValue,
+        onValueChange = {
+            textFieldValue = it
+            onValueChange(textFieldValue.text)
+        },
+        modifier = modifier.fillMaxWidth(),
+        label = { Text(label) },
+        placeholder = { Text(hint) },
+        colors = TextFieldDefaults.outlinedTextFieldColors(
+            focusedLabelColor = MaterialTheme.colorScheme.onPrimary,
+            focusedBorderColor = MaterialTheme.colorScheme.onPrimary,
+            unfocusedBorderColor = MaterialTheme.colorScheme.onPrimary,
+            errorBorderColor = MaterialTheme.colorScheme.error,
+            errorLabelColor = MaterialTheme.colorScheme.error,
+            errorSupportingTextColor = MaterialTheme.colorScheme.error
+        ),
+        singleLine = true,
+        maxLines = 1,
+        shape = MaterialTheme.shapes.small,
+        keyboardOptions = KeyboardOptions.Default.copy(keyboardType = KeyboardType.Number)
     )
 }
